% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIspatial-package.R
\docType{package}
\name{NVIspatial-package}
\alias{NVIspatial}
\alias{NVIspatial-package}
\title{NVIspatial: Tools to facilitate working with Spatial Data at NVI}
\description{
Provides tools to facilitate working with spatial data at NVI. Currently, NVI spatial holds the function 'is_it_in_Norway'. The package will be extended with other tools for handling (Norwegian) spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIspatial}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIspatial/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{petter.hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Authors:
\itemize{
  \item Katharine Rose Dean \email{katharine.rose.dean@vetinst.no} (ORCID = 0000-0003-2262-0385)
}

Other contributors:
\itemize{
  \item Johan Åkerstedt \email{johan.akerstedt@vetinst.no} (ORCID = 0000-0001-7887-6887) [contributor]
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
